pyproject.toml
[project]
name = "havoc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kaggle>=1.7.4.5",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "onnx>=1.19.1",
    "onnxscript>=0.5.4",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "seaborn>=0.13.2",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
]
